requirements.txt
streamlit
ifcopenshell
pandas
plotly
reportlab
anthropic
python-dotenv